# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CertificationCode =
  (typeof CertificationCode)[keyof typeof CertificationCode];

export const CertificationCode = {
  "AZ-900": "AZ-900",
  "AZ-104": "AZ-104",
  "AZ-305": "AZ-305",
} as const;

export type CertificationLevel =
  (typeof CertificationLevel)[keyof typeof CertificationLevel];

export const CertificationLevel = {
  Fundamentals: "Fundamentals",
  Associate: "Associate",
  Expert: "Expert",
} as const;

export interface Certification {
  id: string;
  code: CertificationCode;
  name: string;
  description: string;
  level: CertificationLevel;
  iconColor: string;
  topicAreas: string[];
}

export type ExamConfigCertificationCode =
  (typeof ExamConfigCertificationCode)[keyof typeof ExamConfigCertificationCode];

export const ExamConfigCertificationCode = {
  "AZ-900": "AZ-900",
  "AZ-104": "AZ-104",
  "AZ-305": "AZ-305",
} as const;

export type ExamConfigMode =
  (typeof ExamConfigMode)[keyof typeof ExamConfigMode];

export const ExamConfigMode = {
  practice: "practice",
  exam: "exam",
} as const;

export type ExamConfigDifficulty =
  (typeof ExamConfigDifficulty)[keyof typeof ExamConfigDifficulty];

export const ExamConfigDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  mixed: "mixed",
} as const;

export interface ExamConfig {
  certificationCode: ExamConfigCertificationCode;
  mode: ExamConfigMode;
  difficulty: ExamConfigDifficulty;
  /**
   * @minimum 10
   * @maximum 60
   */
  questionCount: number;
  userId?: string;
}

export type ExamQuestionOptionsItem = {
  key: string;
  text: string;
};

export interface ExamQuestion {
  id: string;
  content: string;
  options: ExamQuestionOptionsItem[];
  correctAnswer: string;
  explanation: string;
  topic: string;
  difficulty: string;
  isAiGenerated: boolean;
  validationScore?: number;
}

export interface ExamSession {
  sessionId: string;
  certificationCode: string;
  mode: string;
  difficulty: string;
  questionCount: number;
  timeLimitMinutes: number;
  questions: ExamQuestion[];
  createdAt: string;
}

export type ExamSubmissionAnswers = { [key: string]: string };

export interface ExamSubmission {
  answers: ExamSubmissionAnswers;
  timeSpentSeconds: number;
}

export type ExamResultQuestionResultsItem = {
  questionId: string;
  correct: boolean;
  userAnswer: string;
  correctAnswer: string;
  explanation: string;
  topic: string;
};

export type ExamResultTopicBreakdownItem = {
  topic: string;
  correct: number;
  total: number;
  percentage: number;
};

export interface ExamResult {
  attemptId: string;
  score: number;
  passed: boolean;
  correctCount: number;
  totalCount: number;
  timeSpentSeconds: number;
  questionResults: ExamResultQuestionResultsItem[];
  topicBreakdown: ExamResultTopicBreakdownItem[];
}

export interface Attempt {
  id: string;
  certificationCode: string;
  mode: string;
  score: number;
  passed: boolean;
  questionCount: number;
  timeSpentSeconds: number;
  completedAt: string;
}

export type AttemptAnalysisStudyPlanItemPriority =
  (typeof AttemptAnalysisStudyPlanItemPriority)[keyof typeof AttemptAnalysisStudyPlanItemPriority];

export const AttemptAnalysisStudyPlanItemPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type AttemptAnalysisStudyPlanItem = {
  topic: string;
  priority: AttemptAnalysisStudyPlanItemPriority;
  recommendation: string;
  resources: string[];
};

export interface AttemptAnalysis {
  overallAssessment: string;
  weakAreas: string[];
  strongAreas: string[];
  studyPlan: AttemptAnalysisStudyPlanItem[];
  nextSteps: string[];
  recommendedNextExam?: string;
  estimatedReadiness: string;
}

export interface AttemptDetail {
  attempt: Attempt;
  result: ExamResult;
  analysis?: AttemptAnalysis;
}

export type GenerateQuestionsRequestCertificationCode =
  (typeof GenerateQuestionsRequestCertificationCode)[keyof typeof GenerateQuestionsRequestCertificationCode];

export const GenerateQuestionsRequestCertificationCode = {
  "AZ-900": "AZ-900",
  "AZ-104": "AZ-104",
  "AZ-305": "AZ-305",
} as const;

export type GenerateQuestionsRequestDifficulty =
  (typeof GenerateQuestionsRequestDifficulty)[keyof typeof GenerateQuestionsRequestDifficulty];

export const GenerateQuestionsRequestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  mixed: "mixed",
} as const;

export interface GenerateQuestionsRequest {
  certificationCode: GenerateQuestionsRequestCertificationCode;
  difficulty: GenerateQuestionsRequestDifficulty;
  /**
   * @minimum 1
   * @maximum 20
   */
  count: number;
  topics?: string[];
}

export interface GenerateQuestionsResponse {
  questions: ExamQuestion[];
  generatedCount: number;
  validatedCount: number;
  rejectedCount: number;
}

export type AnalyzeAttemptRequestTopicBreakdownItem = {
  topic: string;
  correct: number;
  total: number;
  percentage: number;
};

export type AnalyzeAttemptRequestQuestionResultsItem = {
  topic: string;
  correct: boolean;
};

export interface AnalyzeAttemptRequest {
  attemptId: string;
  certificationCode: string;
  score: number;
  topicBreakdown: AnalyzeAttemptRequestTopicBreakdownItem[];
  questionResults: AnalyzeAttemptRequestQuestionResultsItem[];
}