# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AzureCertify AI Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate-questions:
    post:
      operationId: generateQuestions
      tags: [ai]
      summary: Generate AI exam questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionsRequest"
      responses:
        "200":
          description: Generated and validated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateQuestionsResponse"
  /ai/analyze-attempt:
    post:
      operationId: analyzeAttempt
      tags: [ai]
      summary: AI analysis of exam attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeAttemptRequest"
      responses:
        "200":
          description: AI analysis with study recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptAnalysis"
components:
  schemas:
...
    ExamQuestion:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              text:
                type: string
            required: [key, text]
        correctAnswer:
          type: string
        explanation:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        isAiGenerated:
          type: boolean
        validationScore:
          type: number
      required: [id, content, options, correctAnswer, explanation, topic, difficulty, isAiGenerated]
...
    GenerateQuestionsRequest:
      type: object
      properties:
        certificationCode:
          type: string
          enum: [AZ-900, AZ-104, AZ-305]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        count:
          type: integer
          minimum: 1
          maximum: 20
        topics:
          type: array
          items:
            type: string
      required: [certificationCode, difficulty, count]
    GenerateQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ExamQuestion"
        generatedCount:
          type: integer
        validatedCount:
          type: integer
        rejectedCount:
          type: integer
      required: [questions, generatedCount, validatedCount, rejectedCount]
    AnalyzeAttemptRequest:
      type: object
      properties:
        attemptId:
          type: string
        certificationCode:
          type: string
        score:
          type: number
        topicBreakdown:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              correct:
                type: integer
              total:
                type: integer
              percentage:
                type: number
            required: [topic, correct, total, percentage]
        questionResults:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              correct:
                type: boolean
            required: [topic, correct]
      required: [attemptId, certificationCode, score, topicBreakdown, questionResults]
    AttemptAnalysis:
      type: object
      properties:
        overallAssessment:
          type: string
        weakAreas:
          type: array
          items:
            type: string
        strongAreas:
          type: array
          items:
            type: string
        studyPlan:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              priority:
                type: string
                enum: [high, medium, low]
              recommendation:
                type: string
              resources:
                type: array
                items:
                  type: string
            required: [topic, priority, recommendation, resources]
        nextSteps:
          type: array
          items:
            type: string
        recommendedNextExam:
          type: string
        estimatedReadiness:
          type: string
      required: [overallAssessment, weakAreas, strongAreas, studyPlan, nextSteps, estimatedReadiness]