# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get available certifications
 */
export const GetCertificationsResponseItem = zod.object({
  id: zod.string(),
  code: zod.enum(["AZ-900", "AZ-104", "AZ-305"]),
  name: zod.string(),
  description: zod.string(),
  level: zod.enum(["Fundamentals", "Associate", "Expert"]),
  iconColor: zod.string(),
  topicAreas: zod.array(zod.string()),
});
export const GetCertificationsResponse = zod.array(
  GetCertificationsResponseItem,
);

/**
 * @summary Configure and start an exam session
 */
export const configureExamBodyQuestionCountMin = 10;
export const configureExamBodyQuestionCountMax = 60;

export const ConfigureExamBody = zod.object({
  certificationCode: zod.enum(["AZ-900", "AZ-104", "AZ-305"]),
  mode: zod.enum(["practice", "exam"]),
  difficulty: zod.enum(["easy", "medium", "hard", "mixed"]),
  questionCount: zod
    .number()
    .min(configureExamBodyQuestionCountMin)
    .max(configureExamBodyQuestionCountMax),
  userId: zod.string().optional(),
});

export const ConfigureExamResponse = zod.object({
  sessionId: zod.string(),
  certificationCode: zod.string(),
  mode: zod.string(),
  difficulty: zod.string(),
  questionCount: zod.number(),
  timeLimitMinutes: zod.number(),
  questions: zod.array(
    zod.object({
      id: zod.string(),
      content: zod.string(),
      options: zod.array(
        zod.object({
          key: zod.string(),
          text: zod.string(),
        }),
      ),
      correctAnswer: zod.string(),
      explanation: zod.string(),
      topic: zod.string(),
      difficulty: zod.string(),
      isAiGenerated: zod.boolean(),
      validationScore: zod.number().optional(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Get exam session with questions
 */
export const GetExamSessionParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetExamSessionResponse = zod.object({
  sessionId: zod.string(),
  certificationCode: zod.string(),
  mode: zod.string(),
  difficulty: zod.string(),
  questionCount: zod.number(),
  timeLimitMinutes: zod.number(),
  questions: zod.array(
    zod.object({
      id: zod.string(),
      content: zod.string(),
      options: zod.array(
        zod.object({
          key: zod.string(),
          text: zod.string(),
        }),
      ),
      correctAnswer: zod.string(),
      explanation: zod.string(),
      topic: zod.string(),
      difficulty: zod.string(),
      isAiGenerated: zod.boolean(),
      validationScore: zod.number().optional(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Submit exam answers
 */
export const SubmitExamParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const SubmitExamBody = zod.object({
  answers: zod.record(zod.string(), zod.string()),
  timeSpentSeconds: zod.number(),
});

export const SubmitExamResponse = zod.object({
  attemptId: zod.string(),
  score: zod.number(),
  passed: zod.boolean(),
  correctCount: zod.number(),
  totalCount: zod.number(),
  timeSpentSeconds: zod.number(),
  questionResults: zod.array(
    zod.object({
      questionId: zod.string(),
      correct: zod.boolean(),
      userAnswer: zod.string(),
      correctAnswer: zod.string(),
      explanation: zod.string(),
      topic: zod.string(),
    }),
  ),
  topicBreakdown: zod.array(
    zod.object({
      topic: zod.string(),
      correct: zod.number(),
      total: zod.number(),
      percentage: zod.number(),
    }),
  ),
});

/**
 * @summary Get user exam attempts history
 */
export const GetAttemptsResponseItem = zod.object({
  id: zod.string(),
  certificationCode: zod.string(),
  mode: zod.string(),
  score: zod.number(),
  passed: zod.boolean(),
  questionCount: zod.number(),
  timeSpentSeconds: zod.number(),
  completedAt: zod.string(),
});
export const GetAttemptsResponse = zod.array(GetAttemptsResponseItem);

/**
 * @summary Get specific attempt details with AI analysis
 */
export const GetAttemptParams = zod.object({
  attemptId: zod.coerce.string(),
});

export const GetAttemptResponse = zod.object({
  attempt: zod.object({
    id: zod.string(),
    certificationCode: zod.string(),
    mode: zod.string(),
    score: zod.number(),
    passed: zod.boolean(),
    questionCount: zod.number(),
    timeSpentSeconds: zod.number(),
    completedAt: zod.string(),
  }),
  result: zod.object({
    attemptId: zod.string(),
    score: zod.number(),
    passed: zod.boolean(),
    correctCount: zod.number(),
    totalCount: zod.number(),
    timeSpentSeconds: zod.number(),
    questionResults: zod.array(
      zod.object({
        questionId: zod.string(),
        correct: zod.boolean(),
        userAnswer: zod.string(),
        correctAnswer: zod.string(),
        explanation: zod.string(),
        topic: zod.string(),
      }),
    ),
    topicBreakdown: zod.array(
      zod.object({
        topic: zod.string(),
        correct: zod.number(),
        total: zod.number(),
        percentage: zod.number(),
      }),
    ),
  }),
  analysis: zod
    .object({
      overallAssessment: zod.string(),
      weakAreas: zod.array(zod.string()),
      strongAreas: zod.array(zod.string()),
      studyPlan: zod.array(
        zod.object({
          topic: zod.string(),
          priority: zod.enum(["high", "medium", "low"]),
          recommendation: zod.string(),
          resources: zod.array(zod.string()),
        }),
      ),
      nextSteps: zod.array(zod.string()),
      recommendedNextExam: zod.string().optional(),
      estimatedReadiness: zod.string(),
    })
    .optional(),
});

/**
 * @summary Generate AI exam questions
 */
export const generateQuestionsBodyCountMax = 20;

export const GenerateQuestionsBody = zod.object({
  certificationCode: zod.enum(["AZ-900", "AZ-104", "AZ-305"]),
  difficulty: zod.enum(["easy", "medium", "hard", "mixed"]),
  count: zod.number().min(1).max(generateQuestionsBodyCountMax),
  topics: zod.array(zod.string()).optional(),
});

export const GenerateQuestionsResponse = zod.object({
  questions: zod.array(
    zod.object({
      id: zod.string(),
      content: zod.string(),
      options: zod.array(
        zod.object({
          key: zod.string(),
          text: zod.string(),
        }),
      ),
      correctAnswer: zod.string(),
      explanation: zod.string(),
      topic: zod.string(),
      difficulty: zod.string(),
      isAiGenerated: zod.boolean(),
      validationScore: zod.number().optional(),
    }),
  ),
  generatedCount: zod.number(),
  validatedCount: zod.number(),
  rejectedCount: zod.number(),
});

/**
 * @summary AI analysis of exam attempt
 */
export const AnalyzeAttemptBody = zod.object({
  attemptId: zod.string(),
  certificationCode: zod.string(),
  score: zod.number(),
  topicBreakdown: zod.array(
    zod.object({
      topic: zod.string(),
      correct: zod.number(),
      total: zod.number(),
      percentage: zod.number(),
    }),
  ),
  questionResults: zod.array(
    zod.object({
      topic: zod.string(),
      correct: zod.boolean(),
    }),
  ),
});

export const AnalyzeAttemptResponse = zod.object({
  overallAssessment: zod.string(),
  weakAreas: zod.array(zod.string()),
  strongAreas: zod.array(zod.string()),
  studyPlan: zod.array(
    zod.object({
      topic: zod.string(),
      priority: zod.enum(["high", "medium", "low"]),
      recommendation: zod.string(),
      resources: zod.array(zod.string()),
    }),
  ),
  nextSteps: zod.array(zod.string()),
  recommendedNextExam: zod.string().optional(),
  estimatedReadiness: zod.string(),
});