# lib/api-zod/src/generated/types/analyzeAttemptRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeAttemptRequestQuestionResultsItem } from "./analyzeAttemptRequestQuestionResultsItem";
import type { AnalyzeAttemptRequestTopicBreakdownItem } from "./analyzeAttemptRequestTopicBreakdownItem";

export interface AnalyzeAttemptRequest {
  attemptId: string;
  certificationCode: string;
  score: number;
  topicBreakdown: AnalyzeAttemptRequestTopicBreakdownItem[];
  questionResults: AnalyzeAttemptRequestQuestionResultsItem[];
}