# lib/api-zod/src/generated/types/attemptAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptAnalysisStudyPlanItem } from "./attemptAnalysisStudyPlanItem";

export interface AttemptAnalysis {
  overallAssessment: string;
  weakAreas: string[];
  strongAreas: string[];
  studyPlan: AttemptAnalysisStudyPlanItem[];
  nextSteps: string[];
  recommendedNextExam?: string;
  estimatedReadiness: string;
}