# lib/api-zod/src/generated/types/generateQuestionsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateQuestionsRequestCertificationCode } from "./generateQuestionsRequestCertificationCode";
import type { GenerateQuestionsRequestDifficulty } from "./generateQuestionsRequestDifficulty";

export interface GenerateQuestionsRequest {
  certificationCode: GenerateQuestionsRequestCertificationCode;
  difficulty: GenerateQuestionsRequestDifficulty;
  /**
   * @minimum 1
   * @maximum 20
   */
  count: number;
  topics?: string[];
}